+++ lib/views/new_screens/new_welcome_screen.dart
@@ -4,7 +4,7 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:resonate/utils/ui_sizes.dart';
 import 'package:resonate/new_themes/theme_screen.dart';
 import 'package:resonate/views/new_screens/new_email_verification_screen.dart';
-import 'package:resonate/views/widgets/welcome_screen_dialog.dart';
+import 'package:resonate/views/new_widgets/welcome_screen_dialog.dart';
 
 class NewWelcomeScreen extends StatelessWidget {
   const NewWelcomeScreen({super.key});
@@ -143,21 +143,7 @@ class NewWelcomeScreen extends StatelessWidget {
                           foregroundColor:
                               Theme.of(context).colorScheme.onPrimary,
                         ),
-                        onPressed: () {
-                          // Navigator.push(
-                          //   context,
-                          //   MaterialPageRoute(
-                          //     builder: (context) => const ThemeColorsPage(),
-                          //   ),
-                          // );
-
-                          Navigator.push(
-                            context,
-                            MaterialPageRoute(
-                              builder: (context) => NewEmailVerificationScreen(),
-                            ),
-                          );
-                        },
+                        onPressed: () {},
                         icon: const FaIcon(
                           FontAwesomeIcons.github,
                         ),
